+++ PSGSuite/PSGSuite.psm1
@@ -1,18 +1,13 @@
 Param
 (
-    [parameter(Position = 0)]
-    [System.Byte[]]
-    $EncryptionKey = $(if (Get-Command Import-SCRTKey -ErrorAction SilentlyContinue) {
-        Import-SCRTKey
-    }
-    else {
-        $null
-    }),
+    [parameter(Position = 0,ValueFromRemainingArguments = $true)]
+    [AllowNull()]
+    [Byte[]]
+    $EncryptionKey = $null,
     [parameter(Position = 1)]
-    [string]
-    $ConfigName = $null
+    [AllowNull()]
+    [String]
+    $ConfigName
 )
-#Get public and private function definition files.
-$Public = @(Get-ChildItem -Recurse -Path $PSScriptRoot\Public\*.ps1 -ErrorAction SilentlyContinue)
-$Private = @(Get-ChildItem -Recurse -Path $PSScriptRoot\Private\*.ps1 -ErrorAction SilentlyContinue)
 $ModuleRoot = $PSScriptRoot
+New-Variable -Name PSGSuiteKey -Value $EncryptionKey -Scope Global -Force -PassThru