+++ Tests/Unit Tests/Users/PSGSuite.Get-GSUser.Tests.ps1
@@ -1,4 +1,5 @@
 InModuleScope PSGSuite {
+    Write-Verbose "Loading mocked versions of New-GoogleService"
     . ([System.IO.Path]::Combine("$env:BHProjectPath","Tests","Mocks","Users.Mocks.ps1"))
     Describe 'Directory function mock tests' {
         Context 'When a mocked Directory service is created' {