+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 # Changelog
 
+* [2.17.0](#2170)
+* [2.16.1](#2161)
 * [2.16.0](#2160)
 * [2.15.4](#2154)
 * [2.15.3](#2153)
@@ -51,13 +53,23 @@
     * [Functions Removed](#functions-removed)
     * [Functions Aliased](#functions-aliased)
 
+## 2.17.0
+
+* [Issue #102](https://github.com/scrthq/PSGSuite/issues/102)
+  * Fixed: `$EncryptionKey` PSM1 parameter now stores the AES key correctly so SecureStrings are encrypted/decrypted as intended.
+* [Issue #103](https://github.com/scrthq/PSGSuite/issues/103)
+  * Updated: `SendNotificationEmail` parameter on `Add-GSDrivePermission` defaults to false for all User & Group permissions that are not ownership transfers.
+  * Updated: Documentation for `SendNotificationEmail` parameter on `Add-GSDrivePermission` for clarity towards default Google API parameter values.
+* Added: More unit testing for `Get-GSUser`
+* Updated: `psake` build script
+
 ## 2.16.1
 
-- Fixed: Module deployment segment in psake script deploying decompiled/broken module
+* Fixed: Module deployment segment in psake script deploying decompiled/broken module
 
 ## 2.16.0
 
-- Updated: Build script to compile module into a single PSM1 file for cleanliness and loading speed improvements
+* Updated: Build script to compile module into a single PSM1 file for cleanliness and loading speed improvements
 
 ## 2.15.4